pyproject.toml
[project]
name = "hxx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "polars>=1.32.0",
    "pyarrow>=21.0.0",
    "requests>=2.32.4",
    "tenacity>=9.1.2",
    "yfinance>=0.2.65",
    "pandas-ta>=0.3.14b0",
    "setuptools>=80.9.0",
    "importlib-metadata>=8.7.0",
    "akshare>=1.17.26",
    "ccxt>=4.4.98",
    "python-box>=7.3.2",
    "pydantic>=2.11.7",
    "loguru>=0.7.0",
]
